requirements.txt
Flask==3.1.0
scipy==1.15.1
SPARQLWrapper==2.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.1